add permission to emrgencycontact method in controller

+++ app/Http/Controllers/EmergencyContactController.php
@@ -9,29 +9,45 @@ use Illuminate\Support\Facades\Auth;
 
 class EmergencyContactController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $contact = EmrgencyContact::where('user_id', Auth::id())->get();
-        return response()->json($contact);
+        if ($request->user()->can('emrgency.index')) {
+            $contact = EmrgencyContact::where('user_id', Auth::id())->get();
+            return response()->json($contact);
+        } else {
+            return response()->json(['message' => 'شما دسترسی مجاز را ندارید']);
+        }
     }
 
     public function store(EmrgencyContactrequest $request)
     {
-        $contact = EmrgencyContact::create($request->toArray());
-        return response()->json($contact, 201);
+        if ($request->user()->can('emrgency.store')) {
+            $contact = EmrgencyContact::create($request->toArray());
+            return response()->json($contact, 201);
+        } else {
+            return response()->json(['message' => 'شما دسترسی مجاز را ندارید']);
+        }
     }
 
     public function update(EmrgencyContactrequest $request, $id)
     {
-        $contact = EmrgencyContact::findOrFail($id);
-        $contact->update($request->only(['name', 'phone']));
-        return response()->json(['message' => 'شماره اضطراری با موفقیت به روز رسانی شد',$contact]);
+        if ($request->user()->can('emrgency.update')) {
+            $contact = EmrgencyContact::findOrFail($id);
+            $contact->update($request->only(['name', 'phone']));
+            return response()->json(['message' => 'شماره اضطراری با موفقیت به روز رسانی شد', $contact]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی مجاز را ندارید']);
+        }
     }
 
     public function destroy(Request $request, $id)
     {
-        $contact = EmrgencyContact::findOrFail($id);
-        $contact->delete();
-        return response()->json(['message' => 'تماس اضطراری با موفقیت حذف شد',]);
+        if ($request->user()->can('emrgency.destroy')) {
+            $contact = EmrgencyContact::findOrFail($id);
+            $contact->delete();
+            return response()->json(['message' => 'تماس اضطراری با موفقیت حذف شد',]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی مجاز را ندارید']);
+        }
     }
 }
